requirements.txt
# requirements.txt

streamlit==1.34.0
langchain==0.1.12
langchain-community==0.0.27
langchain-groq==0.1.0
pypdf==4.2.0
faiss-cpu==1.8.0
sentence-transformers==2.5.1
python-dotenv==1.0.1